perf(fretboard): memoize per-note rendering so unchanged notes skip re-render
